// src/schema.ts
import { z } from "@zod/zod";

/**
 * A Zod Schema for the deno.json config file.
 *
 * Based off https://github.com/denoland/deno/blob/main/cli/schemas/config-file.v1.json
 */
export type DenoConfigFile = z.infer<typeof DenoConfigFile>;
export const DenoConfigFile = z.object({
  name: z.string().optional(),
  version: z.string().optional(),
  exports: z.union([z.string(), z.record(z.string(), z.string())]).optional(),
  compilerOptions: z.object({
    allowImportingTsExtensions: z.boolean().optional(),
    allowJs: z.boolean().optional(),
    allowUnreachableCode: z.boolean().optional(),
    allowUnusedLabels: z.boolean().optional(),
    checkJs: z.boolean().optional(),
    experimentalDecorators: z.boolean().optional(),
    jsx: z.enum([
      "react",
      "react-jsx",
      "react-jsxdev",
      "react-native",
      "preserve",
      "precompile",
    ]).optional(),
    jsxFactory: z.string().optional(),
    jsxFragmentFactory: z.string().optional(),
    jsxImportSource: z.string().optional(),
    keyofStringsOnly: z.boolean().optional(),
    lib: z.array(z.string()).optional(),
    moduleResolution: z.enum(["node", "classic"]).optional(),
    noEmitOnError: z.boolean().optional(),
    noImplicitAny: z.boolean().optional(),
    noImplicitOverride: z.boolean().optional(),
    noImplicitReturns: z.boolean().optional(),
    noImplicitThis: z.boolean().optional(),
    noImplicitUseStrict: z.boolean().optional(),
    noStrictGenericChecks: z.boolean().optional(),
    noUncheckedIndexedAccess: z.boolean().optional(),
    strict: z.boolean().optional(),
    strictBindCallApply: z.boolean().optional(),
    strictFunctionTypes: z.boolean().optional(),
    strictNullChecks: z.boolean().optional(),
    strictPropertyInitialization: z.boolean().optional(),
    suppressExcessPropertyErrors: z.boolean().optional(),
    suppressImplicitAnyIndexErrors: z.boolean().optional(),
    target: z.enum([
      "es3",
      "es5",
      "es6",
      "es2015",
      "es2016",
      "es2017",
      "es2018",
      "es2019",
      "es2020",
      "es2021",
      "es2022",
      "esnext",
    ]).optional(),
    useDefineForClassFields: z.boolean().optional(),
    useUnknownInCatchVariables: z.boolean().optional(),
  }).optional(),
  lint: z.object({
    include: z.array(z.string()).optional(),
    exclude: z.array(z.string()).optional(),
    rules: z.object({
      tags: z.array(z.string()).optional(),
      exclude: z.array(z.string()).optional(),
      include: z.array(z.string()).optional(),
    }).optional(),
    report: z.enum(["pretty", "json", "compact"]).optional(),
  }).optional(),
  fmt: z.object({
    include: z.array(z.string()).optional(),
    exclude: z.array(z.string()).optional(),
    useTabs: z.boolean().optional(),
    lineWidth: z.number().optional(),
    indentWidth: z.number().optional(),
    singleQuote: z.boolean().optional(),
    proseWrap: z.enum(["always", "never", "preserve"]).optional(),
    semiColons: z.boolean().optional(),
  }).optional(),
  tasks: z.record(z.string(), z.string()).optional(),
  test: z.object({
    include: z.array(z.string()).optional(),
    exclude: z.array(z.string()).optional(),
  }).optional(),
  bench: z.object({
    include: z.array(z.string()).optional(),
    exclude: z.array(z.string()).optional(),
  }).optional(),
  importMap: z.string().optional(),
  imports: z.record(z.string(), z.string()).optional(),
  scopes: z.record(z.string(), z.record(z.string(), z.string())).optional(),
  nodeModulesDir: z.union([z.boolean(), z.string()]).optional(),
  vendor: z.boolean().optional(),
  lock: z.union([z.boolean(), z.string()]).optional(),
  workspace: z.object({
    members: z.array(z.string()).optional(),
  }).optional(),
  publish: z.object({
    include: z.array(z.string()).optional(),
    exclude: z.array(z.string()).optional(),
  }).optional(),
  unstable: z.union([z.array(z.string()), z.boolean()]).optional(),
});